Backend/app/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==1.1.2
tweepy==3.9.0
requests==2.22.0
pyjwt==1.7.1
pandas==1.0.1
numpy==1.18.1
nltk==3.5
mysql-connector-python==8.0.22
schedule==0.6.0
scikit-learn==0.23.2
werkzeug==1.0.0
textblob==0.15.3
gitpython==3.0.7
flask-limiter==1.4
flask-cors==3.0.9
joblib==1.0.0
mysql
waitress
flask-csp
mysql-connector
jsonschema
python-dotenv